finished error handling in client side, 	-error when crud product 	-error when invalid login finished error handling in server side, finished create login and regiter for normal user finished create regiter for admin finished create navbar components

+++ ecommerce-client/src/components/itemlist.vue
@@ -11,35 +11,57 @@
         <router-link :to="{ name: 'Detail', params: { id: this.data.id }}">
           <button class="btn"><i class="fa fa-bars"></i></button>
         </router-link>
-        <button class="btn" @click="deleteId()"> <i class="fa fa-trash"></i></button>
+        <button v-if="level==='admin'" class="btn" @click="deleteId()">
+          <i class="fa fa-trash"></i></button>
       </div>
   </div>
 </div>
 </template>
 <script>
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 export default {
   props: ['data'],
   data() {
     return {
       id: this.data.id,
+      level: localStorage.getItem('level'),
     };
   },
   methods: {
     deleteId() {
-      axios({
-        method: 'delete',
-        url: `${this.$store.state.axiosUrl}/product/${this.id}`,
-        headers: { token: localStorage.getItem('token') },
-      })
-        .then((data) => {
-          this.$store.dispatch('deleteData', data.id);
-        })
-        .catch((err) => {
-          console.log(err);
-        });
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!',
+      }).then((result) => {
+        if (result.value) {
+          axios({
+            method: 'delete',
+            url: `${this.$store.state.axiosUrl}/product/${this.id}`,
+            headers: { token: localStorage.getItem('token') },
+          })
+            .then((data) => {
+              this.$store.dispatch('deleteData', data.id);
+              this.$store.dispatch('getAllItem');
+            })
+            .catch((err) => {
+              console.log(err);
+            });
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success',
+          );
+        }
+      });
     },
+
   },
 };
 </script>
